moving api-lookups into previewApiLookup folder, get rid off utils folder
