+++ packages/destination-actions/src/destinations/engage/sendgrid/previewApiLookup/api-lookups.ts
@@ -4,7 +4,7 @@ import { RequestClient, RequestOptions } from '@segment/actions-core'
 import { Logger, StatsClient } from '@segment/actions-core/destination-kit'
 import type { Settings } from '../generated-types'
 import { Liquid as LiquidJs } from 'liquidjs'
-import { Profile } from './types'
+import { Profile } from '../Profile'
 
 const Liquid = new LiquidJs()
