+++ packages/destination-actions/src/destinations/engage/sendgrid/sendEmail/actionDefinition.ts
@@ -1,7 +1,7 @@
 import { ActionDefinition, IntegrationError } from '@segment/actions-core'
 import type { Settings } from '../generated-types'
 import type { Payload } from './generated-types'
-import { apiLookupActionFields } from '../utils/api-lookups'
+import { apiLookupActionFields } from '../previewApiLookup/api-lookups'
 import { attemptEmailDelivery } from './attemptEmailDelivery'
 
 export const actionDefinition: ActionDefinition<Settings, Payload> = {
