+++ packages/destination-actions/src/destinations/engage/sendgrid/sendEmail/attemptEmailDelivery.ts
@@ -2,8 +2,8 @@ import { IntegrationError, RequestClient, RequestOptions } from '@segment/action
 import type { Settings } from '../generated-types'
 import type { Payload } from './generated-types'
 import { Logger, StatsClient } from '@segment/actions-core/destination-kit'
-import { apiLookupLiquidKey, performApiLookups } from '../utils/api-lookups'
-import { Profile } from '../utils/types'
+import { apiLookupLiquidKey, performApiLookups } from '../previewApiLookup/api-lookups'
+import { Profile } from '../Profile'
 import { insertEmailPreviewText } from './insertEmailPreviewText'
 import { insertUnsubscribeLinks } from './insertUnsubscribeLinks'
 import { fetchProfileTraits } from './fetchProfileTraits'
